refactor(internal/database):rename metadata test package to test_impl
