+++ internal/database/metadata/postgres/entity_test.go
@@ -3,21 +3,21 @@ package postgres_test
 import (
 	"testing"
 
-	"github.com/oom-ai/oomstore/internal/database/metadata/test"
+	"github.com/oom-ai/oomstore/internal/database/metadata/test_impl"
 )
 
 func TestCreateEntity(t *testing.T) {
-	test.TestCreateEntity(t, prepareStore)
+	test_impl.TestCreateEntity(t, prepareStore)
 }
 
 func TestGetEntity(t *testing.T) {
-	test.TestGetEntity(t, prepareStore)
+	test_impl.TestGetEntity(t, prepareStore)
 }
 
 func TestUpdateEntity(t *testing.T) {
-	test.TestUpdateEntity(t, prepareStore)
+	test_impl.TestUpdateEntity(t, prepareStore)
 }
 
 func TestListEntity(t *testing.T) {
-	test.TestListEntity(t, prepareStore)
+	test_impl.TestListEntity(t, prepareStore)
 }
