+++ internal/database/metadata/postgres/feature_group_test.go
@@ -3,21 +3,21 @@ package postgres_test
 import (
 	"testing"
 
-	"github.com/oom-ai/oomstore/internal/database/metadata/test"
+	"github.com/oom-ai/oomstore/internal/database/metadata/test_impl"
 )
 
 func TestGetFeatureGroup(t *testing.T) {
-	test.TestGetFeatureGroup(t, prepareStore)
+	test_impl.TestGetFeatureGroup(t, prepareStore)
 }
 
 func TestListFeatureGroup(t *testing.T) {
-	test.TestListFeatureGroup(t, prepareStore)
+	test_impl.TestListFeatureGroup(t, prepareStore)
 }
 
 func TestCreateFeatureGroup(t *testing.T) {
-	test.TestCreateFeatureGroup(t, prepareStore)
+	test_impl.TestCreateFeatureGroup(t, prepareStore)
 }
 
 func TestUpdateFeatureGroup(t *testing.T) {
-	test.TestUpdateFeatureGroup(t, prepareStore)
+	test_impl.TestUpdateFeatureGroup(t, prepareStore)
 }
