+++ internal/database/metadata/postgres/revision_test.go
@@ -3,25 +3,25 @@ package postgres_test
 import (
 	"testing"
 
-	"github.com/oom-ai/oomstore/internal/database/metadata/test"
+	"github.com/oom-ai/oomstore/internal/database/metadata/test_impl"
 )
 
 func TestCreateRevision(t *testing.T) {
-	test.TestCreateRevision(t, prepareStore)
+	test_impl.TestCreateRevision(t, prepareStore)
 }
 
 func TestUpdateRevision(t *testing.T) {
-	test.TestUpdateRevision(t, prepareStore)
+	test_impl.TestUpdateRevision(t, prepareStore)
 }
 
 func TestGetRevision(t *testing.T) {
-	test.TestGetRevision(t, prepareStore)
+	test_impl.TestGetRevision(t, prepareStore)
 }
 
 func TestGetRevisionBy(t *testing.T) {
-	test.TestGetRevisionBy(t, prepareStore)
+	test_impl.TestGetRevisionBy(t, prepareStore)
 }
 
 func TestListRevision(t *testing.T) {
-	test.TestListRevision(t, prepareStore)
+	test_impl.TestListRevision(t, prepareStore)
 }
