+++ internal/database/metadata/test_impl/helper.go
@@ -1,4 +1,4 @@
-package test
+package test_impl
 
 func boolPtr(i bool) *bool {
 	return &i
