+++ internal/database/metadata/test_impl/revision.go
@@ -1,4 +1,4 @@
-package test
+package test_impl
 
 import (
 	"context"
